+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 
 ## Added
 
+- **Supervisor scheduling** - Schedule an actor spawning.
 - **Drainer** - A special task-agent in the superagent crate to drain events sources.
 - **ToRecipient** - A trait to produce a recipient for events.
 - **Ping** - A ping extension has been added.